refactor: catalog managers
