+++ src/catalog/Cargo.toml
@@ -32,6 +32,7 @@ meta-client = { workspace = true }
 metrics.workspace = true
 moka = { version = "0.11", features = ["future"] }
 parking_lot = "0.12"
+partition.workspace = true
 regex.workspace = true
 serde.workspace = true
 serde_json = "1.0"
