+++ src/catalog/src/lib.rs
@@ -31,9 +31,9 @@ use crate::error::Result;
 
 pub mod error;
 pub mod information_schema;
+pub mod kvbackend;
 pub mod memory;
 mod metrics;
-pub mod remote;
 pub mod system;
 pub mod table_source;
