+++ src/catalog/src/memory.rs
@@ -11,21 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#![feature(assert_matches)]
 
-pub mod dist_table;
-pub mod engine;
-pub mod error;
-pub mod metadata;
-pub mod predicate;
-pub mod requests;
-pub mod stats;
-pub mod table;
-pub mod test_util;
-pub mod thin_table;
+pub mod manager;
 
-pub use store_api::storage::RegionStat;
-
-pub use crate::error::{Error, Result};
-pub use crate::stats::{ColumnStatistics, TableStatistics};
-pub use crate::table::{Table, TableRef};
+pub use manager::{new_memory_catalog_manager, MemoryCatalogManager};
