+++ src/cmd/src/standalone.rs
@@ -14,7 +14,7 @@
 
 use std::sync::Arc;
 
-use catalog::remote::DummyKvCacheInvalidator;
+use catalog::kvbackend::{DummyKvCacheInvalidator, KvBackendCatalogManager};
 use catalog::CatalogManagerRef;
 use clap::Parser;
 use common_base::Plugins;
@@ -26,7 +26,6 @@ use common_telemetry::logging::LoggingOptions;
 use datanode::datanode::builder::DatanodeBuilder;
 use datanode::datanode::{Datanode, DatanodeOptions, ProcedureConfig, StorageConfig};
 use datanode::region_server::RegionServer;
-use frontend::catalog::FrontendCatalogManager;
 use frontend::frontend::FrontendOptions;
 use frontend::instance::{FrontendInstance, Instance as FeInstance, StandaloneDatanodeManager};
 use frontend::service_config::{
@@ -313,7 +312,7 @@ impl StartCommand {
             .context(StartDatanodeSnafu)?;
         let region_server = datanode.region_server();
 
-        let catalog_manager = FrontendCatalogManager::new(
+        let catalog_manager = KvBackendCatalogManager::new(
             kv_store.clone(),
             Arc::new(DummyKvCacheInvalidator),
             Arc::new(StandaloneDatanodeManager(region_server.clone())),
