+++ src/query/src/query_engine.rs
@@ -138,7 +138,7 @@ mod tests {
 
     #[test]
     fn test_query_engine_factory() {
-        let catalog_list = catalog::local::new_memory_catalog_manager().unwrap();
+        let catalog_list = catalog::memory::new_memory_catalog_manager().unwrap();
         let factory = QueryEngineFactory::new(catalog_list, None, false);
 
         let engine = factory.query_engine();
